enable merge queues
