+++ .github/actions/check-create-redwood-app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "only_doc_changes",
+  "name": "check-test-project-fixture",
   "private": true,
   "dependencies": {
     "@actions/core": "1.10.0",
