+++ .github/actions/check-test-project-fixture/action.yml
@@ -1,9 +1,11 @@
 name: Check test project fixture
 description: Determines if the test project fixture should be rebuilt
+
 runs:
   # `node18` isn't supported yet
   using: node16
-  main: check_test_project_fixture.mjs
+  main: checkTestProjectFixture.mjs
+
 inputs:
   labels:
     required: true
