+++ .github/actions/only-doc-changes/action.yml
@@ -1,9 +1,11 @@
-name: Docs
+name: Only doc changes
 description: Determines if the PR only changes docs
+
 outputs:
   only-doc-changes:
     description: If the PR only changes docs
+
 runs:
   # `node18` isn't supported yet
   using: node16
-  main: only_doc_changes.mjs
+  main: onlyDocChanges.mjs
