+++ .github/actions/only-doc-changes/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "check_test_project_fixture",
+  "name": "only-doc-changes",
   "private": true,
   "dependencies": {
     "@actions/core": "1.10.0",
