+++ .github/workflows/codeql-analysis.yml
@@ -19,6 +19,7 @@ on:
     branches: [ main ]
   schedule:
     - cron: '42 5 * * 3'
+  merge_group:
 
 # Cancel in-progress runs of this workflow.
 # See https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow.
