+++ .github/workflows/require-milestone.yml
@@ -3,6 +3,7 @@ name: 🚩 Require milestone
 on:
   pull_request:
     types: [opened, synchronize, reopened, milestoned, demilestoned]
+  merge_group:
 
 # Cancel in-progress runs of this workflow.
 # See https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow.
@@ -23,5 +24,5 @@ jobs:
         with:
           node-version: 18
 
-      - name: ✅ Check milestone
+      - name: 🚩 Require milestone
         uses: ./.github/actions/require-milestone
