--- .github/workflows/update-all-contributors.yml
@@ -1,32 +1,0 @@
-name: Update all contributors
-
-on:
-  schedule:
-    # This runs once a week: https://crontab.guru/once-a-week.
-    # * is a special character in YAML so you have to quote this string.
-    - cron: '0 0 * * 0'
-  workflow_dispatch:
-
-# Cancel in-progress runs of this workflow.
-# See https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow.
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  update-all-contributors:
-    name: Update all contributors
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          token: ${{ secrets.JTOAR_TOKEN }}
-
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 18
-
-      - run: yarn install
-
-      - name: Update all contributors
-        uses: ./.github/actions/update_all_contributors
